backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "clerk-backend-api>=2.1.0",
    "fastapi>=0.115.13",
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
    "svix>=1.67.0",
    "uuid>=1.30",
    "uvicorn>=0.34.3",
]
